api/requirements.txt
Flask==0.10.1
py3k-bcrypt==0.3
pymongo==2.7.1
pyzmq==14.3.1
pytest==2.5.2
voluptuous==0.8.5
pylibmc>=1.3.0
python3-memcached==1.51
